+++ src/main/java/roomescape/auth/AuthConfig.java
@@ -1,0 +1,13 @@
+package roomescape.auth;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class AuthConfig {
+
+    @Bean
+    JwtTokenManager jwtTokenManager() {
+        return new JwtTokenManager();
+    }
+}
